Add support to publish.sh to be able to add patch

+++ publish.sh
@@ -11,15 +11,44 @@ get_commit() {
 printarr() { declare -n __p="$1"; for k in "${!__p[@]}"; do printf "%s=%s\n" "$k" "${__p[$k]}" ; done ;  }
 
 update_tag() {
+    echo update_tag for $1
     for k in "${!repo_tag[@]}"; do
-        repo_info="${repo_tag[$k]}"
+        local repo_info="${repo_tag[$k]}"
         if [[ "$repo_info" =~ ^tag: ]]; then
-            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type == "git" and .url == "https://github.com/fcitx/'$k'")) .tag = "'${repo_info/tag:/}'"' $1
-            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type == "git" and .url == "https://github.com/fcitx/'$k'")) |= del(.branch)' $1
+            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type? == "git" and .url? == "https://github.com/fcitx/'$k'")) .tag = "'${repo_info/tag:/}'"' $1
+            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type? == "git" and .url? == "https://github.com/fcitx/'$k'")) |= del(.branch)' $1
         elif [[ "$repo_info" =~ ^commit: ]]; then
-            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type == "git" and .url == "https://github.com/fcitx/'$k'")) .commit = "'${repo_info/commit:/}'"' $1
-            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type == "git" and .url == "https://github.com/fcitx/'$k'")) |= del(.branch)' $1
+            echo ABCD
+            echo yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type? == "git" and .url? == "https://github.com/fcitx/'$k'")) .commit = "'${repo_info/commit:/}'"' $1
+            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type? == "git" and .url? == "https://github.com/fcitx/'$k'")) .commit = "'${repo_info/commit:/}'"' $1
+            echo DEFG
+            echo yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type? == "git" and .url? == "https://github.com/fcitx/'$k'")) |= del(.branch)' $1
+            yq -y -i '(.. |select(.sources?) | .sources[]? | select(.type? == "git" and .url? == "https://github.com/fcitx/'$k'")) |= del(.branch)' $1
         fi
+    done
+}
+
+update_cherry_pick() {
+    rm -f $GIT_REPO/cp-*.patch
+
+    if [[ ! -f cherry-picks/$REPO ]]; then
+        return
+    fi
+
+    for commit in `cat cherry-picks/$REPO`; do
+        wget https://github.com/fcitx/$REPO/commit/$commit.patch -O $GIT_REPO/cp-$commit.patch
+        yq -y -i '(.. |select(.modules?) | .modules[]? | select(.name? == "fcitx5")).sources += [{"type": "patch", "path": "'cp-$commit.patch'", "use-git": true}] ' $1
+    done
+}
+
+populate_modules() {
+    local basedir=$(dirname $1)
+    for module in $(yq -r '.modules[]? | select(type == "string" and endswith(".yaml"))' $1); do
+        local moduledir=$(dirname $module)
+        mkdir -p $GIT_REPO/$basedir/$moduledir
+        cp $basedir/$module $GIT_REPO/$basedir/$module
+        update_tag $GIT_REPO/$basedir/$module
+        populate_modules $basedir/$module
     done
 }
 
@@ -64,11 +93,9 @@ fi
 
 cp $PACKAGE.yaml $GIT_REPO/$PACKAGE.yaml
 
-for module in $(yq -r '.modules[] | select(type == "string" and startswith("modules/"))' $GIT_REPO/$PACKAGE.yaml); do
-    mkdir -p $GIT_REPO/modules
-    cp $module $GIT_REPO/$module
-    update_tag $GIT_REPO/$module
-done
+rm -rf $GIT_REPO/modules
+
+populate_modules $PACKAGE.yaml
 
 yq -y -i '.branch = "stable"' $GIT_REPO/$PACKAGE.yaml
 if [[ "$PACKAGE" =~ .*Addon.* ]]; then
@@ -78,7 +105,12 @@ else
     yq -y -i '."add-extensions"."org.fcitx.Fcitx5.Addon".version = "stable"' $GIT_REPO/$PACKAGE.yaml
 fi
 
+if [[ $REPO == "mozc" ]]; then
+    cp mozc-deps.yaml zip-code.patch $GIT_REPO/
+fi
+
 update_tag $GIT_REPO/$PACKAGE.yaml
+update_cherry_pick $GIT_REPO/$PACKAGE.yaml
 
 cd $GIT_REPO
 LABEL=${repo_tag[$REPO]/:/-}
@@ -88,7 +120,7 @@ if [[ "$2" == "new" ]]; then
     git commit -a -m "Add $PACKAGE"
 fi
 
-if [[ "$2" != "new" ]]; then
+if [[ "$2" != "new" ]] && [[ "$2" != dry ]]; then
     git checkout -b pr-$LABEL
     git commit -a -m "Update $PACKAGE"
     git push origin --force pr-$LABEL
